+++ Projects/Feature/AppGroupFeature/Interface/Sources/Constants.swift
@@ -1,0 +1,12 @@
+//
+//  Constants.swift
+//  FeatureAppGroupFeatureInterface
+//
+//  Created by Greem on 7/31/25.
+//
+
+import Foundation
+
+enum Constants {
+    
+}
