+++ Projects/Feature/Onboarding/Interface/Sources/Constants.swift
@@ -1,0 +1,18 @@
+//
+//  Constants.swift
+//  FeatureOnboardingInterface
+//
+//  Created by Greem on 7/31/25.
+//
+
+import Foundation
+
+enum Constants {
+    // MARK: - Privacy Policy & Terms of Service URLs
+    enum WebURLLinks {
+        /// 개인정보 처리 방침 URL
+        static let privacyPolicy = "https://www.notion.so/ahnsh/Brake-223b76e3040280438cc7ec812de68c0f?source=copy_link"
+        /// 이용약관 URL
+        static let termsOfService = "https://www.naver.com"
+    }
+}
