+++ Projects/Feature/Onboarding/Interface/Sources/View/Helper/WebView.swift
@@ -11,7 +11,9 @@ public struct WebView: UIViewRepresentable {
 
     public func makeUIView(context: Context) -> WKWebView {
         let webview = WKWebView()
-        webview.backgroundColor = .white
+        webview.backgroundColor = .black
+        webview.scrollView.backgroundColor = .black
+        webview.scrollView.contentInsetAdjustmentBehavior = .never
         webview.load(URLRequest(url: url))
         return webview
     }
