+++ Projects/Feature/Onboarding/Interface/Sources/View/Login/LoginView.swift
@@ -8,10 +8,10 @@ import SwiftUI
 import Domain
 import WebKit
 import SharedDesignSystem
+
 public struct LoginView: View {
     @Environment(LogInViewModel.self) var logInViewModel
-    @State private var wkURL: URL? = URL(string: "https://www.naver.com")
-    @State private var shouldShowWebView: Bool = false
+    @State private var linkInfoItem: LinkInfoItem? = nil
     @State private var kakaoLogInShow: Bool = false
     @State private var loading: Bool = false
     
@@ -28,8 +28,15 @@ public struct LoginView: View {
                 } appleLogInButtonTapped: {
                     logInViewModel.appleLogInBtnTapped()
                 } privacyInfoButtonTapped: {
-                    wkURL = URL(string: "https://www.naver.com")
-                    shouldShowWebView = true
+                    self.linkInfoItem = LinkInfoItem(
+                        title: "brake.site",
+                        url: URL(string: Constants.WebURLLinks.privacyPolicy)
+                        )
+                } termsOfServiceButtonTapped: {
+                    self.linkInfoItem = LinkInfoItem(
+                        title: "brake.site",
+                        url: URL(string: Constants.WebURLLinks.termsOfService)
+                        )
                 }
                 .padding([.horizontal, .bottom], 16)
             }
@@ -64,33 +71,37 @@ public struct LoginView: View {
                 }
             }
         )
-        .fullScreenCover(isPresented: $shouldShowWebView) {
+        .fullScreenCover(item: self.$linkInfoItem, content: { linkInfoItem in
             NavigationView {
-                if let url = wkURL {
+                if let url = linkInfoItem.url {
                     WebView(url: url)
-                        .frame(maxWidth: .infinity, maxHeight: .infinity)
-                        .navigationTitle("약관")
-                        .navigationBarTitleDisplayMode(.inline)
-                        .navigationBarItems(
-                            trailing: Button("닫기") {
-                                shouldShowWebView = false
-                            }
-                        )
+                    .ignoresSafeArea(.container, edges: .bottom)
+                    .navigationTitle(linkInfoItem.title)
+                    .navigationBarTitleDisplayMode(.inline)
+                    .toolbar(content: {
+                        ToolbarItem(placement: .topBarLeading) {
+                            Button("완료") { self.linkInfoItem = nil }
+                        }
+                    })
                 } else {
                     Text("잘못된 URL입니다")
                         .navigationTitle("오류")
                         .navigationBarTitleDisplayMode(.inline)
-                        .navigationBarItems(
-                            trailing: Button("닫기") {
-                                shouldShowWebView = false
+                        .toolbar(content: {
+                            ToolbarItem(placement: .topBarLeading) {
+                                Button("완료") { self.linkInfoItem = nil }
                             }
-                        )
+                        })
                 }
             }
-        }
+        })
     }
-    
 }
 
+private struct LinkInfoItem: Identifiable {
+    var id: String { title }
+    let title: String
+    let url: URL?
+}
